Placing multiple vtk files into one render
Updated readgrid.py to enable it to place 2 renders on the screen at the same time. This enables me to place the heart model and the vasucalar structure on the smae render window so I can see results. \n\n Also created files to manipulate the oxygenation map of VascusSynth to get it to generate roots to specific points, however this is not complete